Prevent TabFrame being visible
With Phoenix, the TabFrame is unintentionally visible in the top left corner (visible as horizontal and vertical lines with a length of 20px each that are lighter than the border color if tabs are at the top, or as a 20x20px square if tabs are at the bottom). Setting its size to 0, 0 prevents this.