Rework fix for #define type extraction
A user running 32-bit Windows reported a problem with wx.VSCROLL after my
changes in aa0be27.  It is likely some integer conversion is not happening
correctly between C++ and Python.  To resolve this, convert the #defines that
are using standard WX 32-bit values back to int and instead insert casts into
the SIP generated cpp code to resolve the C++11 compilation errors.