Fix stc compilation under SIP 4.18
The stc code doesn't compile under SIP 4.18.  This appears to be due to changes
in the cast_xyz function generation, but it exposed a problem in the Phoenix
implementation.  The wx interface headers say that wxStyledTextCtrl inherits
from wxTextEntry, but it actually inherits from wxTextCtrlIface.  This now
(probably rightly so) doesn't compile with SIP 4.18.  The solution is to
essentially copy the methods from wxTextEntry and wxTextCtrl, as is done for
wxSearchCtrl and wxRichTextCtrl.