Created options for build.py for ensuring precense of required packages
Note: Does not work for Ubuntu16.10, as gstreamer 0.10 is not available. This may be remedied by additional repositories. 
Attempts to ensure that all necessary packages to build Phoenix are present. Packages are separated into two main groups, required and recommended. If a required package is missing when checkdeps is used then the build will exit, if instdeps is used then it will prevent moving forward until it is installed.
A yaml manifest file was created for translating necessary packages to distribution specific names., an additional field "other" is provided for listing packages specific to on distribution.