Update AuiToolBar orientation in AuiManager.DoUpdate
Fix an issue in wx.lib.agw.aui.AuiManager where the orientation of
an AuiToolBar would not be updated when calling LoadPerspective.

This means that if the orientation of the toolbar can be wrong after
calling LoadPerspective.  Example:  The toolbar has been moved by
the user to the right side of the aui widget so that it is oriented
vertically.  SavePerspective is called to save the arrangement of
the aui panels before the application exits.  When the application
starts up the toolbar is created in the default position and is
oriented horizontally.  LoadPerspective is then called to restore
the saved panel arrangement.  The position of the toolbar will be
restored but the orientation will be still horizonal after the
call, which is probably not what the user expects.

A call to SwitchToolBarOrientation in DoUpdate (which is called
by LoadPerspective) has been commented out since the AuiManager
class was imported into wxPython way back in 2009 so I have been
unable to find out why it was commented out.  Removing the comment
makes the above scenario work properly.  I have done some quick
testing of the change and it does not seem to break anything.

Fixes #917