Add additional key handling to customtreectrl
Allow use of numpad +. -. and * keys as well as the Windows menu key in
addition to regular keys.
Fix wrong method call to self.ExpandAll(), now self.ExpandAllChildren().
Note: The code still compares against ord('+') and ord('*') which isn't
valid for wx keycodes and will never evaluate properly.
If the user presses the plus '+' button, on most keyboard this will
result in an ord('=') keycode with the SHIFT modifier set.