Guard against unlinked items in CustomTreeCtrl
The GetNextSibling() and GetPrevSibling() methods in CustomTreeCtrl
will throw an exception if the item has been deleted from the tree.
This happens whenever an item is deleted but the selection isn't
updated. The exception will be thrown on the next key press. Can be
observed easily in the HyperTreeList demo.