Merge item hiding to CustomTreeCtrl
Move item hiding from HyperTreeList to CustomTreeCtrl so that both
trees can have the feature.

Add a HideItemWindows() method to hide windows only for a specific item
and its children. This is much faster than calling HideWindows().

Add context menu options to the CustomTreeCtrl demo so that Hiding
can be tested.