Improve CustomTreeCtrl update speed when frozen
If CustomTreeCtrl or HyperTreeList are frozen, skip the CalculateSize(),
CalculatePositions(), and AdjustMyScrollBars() methods.
These methods are quite slow and are called repeatedly whenever changes
are made to the tree. Skipping constant re-calculating allows rapid
changes to be made to the tree while it is frozen. The self._dirty flag
is set to indicate recalculation is necessary in the next idle handler.