Recalculate CustomTreeCtrl on multiline text
In CustomTreeCtrl and HyperTreeList the SetItemText method recalculates
the size of the item row in which the text was changed. If the number
of lines in the text changes the entire tree must be recalculated
as all items below must be shifted to compensate. This change attempts
to recalculate the entire tree only when necessary (when number of
lines change) in order not to slow down existing code that performs a
lot of SetItemText() operations on an unfrozen tree.

Add a "Change Item Text" option to the CustomTreeCtrl and HyperTreeList
demos so that this can be tested. Before this fix, changing the number
of lines of text in an item would leave gaps or create overlaps
between items in the tree.