Fix HyperTreeList ellipsized text alignment
If the TR_ELLIPSIZE_LONG_ITEMS flag is set the HyperTreeList will not
center or right-align text properly if it is too long since the size
calculation of the text is performed before the call to ChopText().
This change flips the calculation to be after ChopText().

Can be seen in the HyperTreeList demo on columns with long text (will
have enter long text manually). Make sure the TR_ELLIPSIZE_LONG_ITEMS
checkbox is enabled and change the alignment to wx.CENTER. Adjust the
column width and text placement will be too far to the right.