Fix HyperTreeList column width calculation
The calculation for item width was not adding in the width of the
item's control (checkbox, radiobutton) or margin.

This can be seen in the demo by deleting 'Item 1' and expanding the
other items to reveal items with checkboxes and radiobuttons. Then
auto-size the tree column by double-clicking on the column separator.
The column will auto-size too small, cutting off text for the items
with radiobuttons or checkboxes.