Fix HyperTreeList exception on item drag in GTK
The call to wx.DragImage.BeginDrag() captures the mouse. At this point
the tree has already captured the mouse. This is not a problem in MSW
or MacOS but in GTK it throws a wx._core.wxAssertionError:
  C++ assertion "!wxMouseCapture::IsInCaptureStack(this)" failed at
  /tmp/pip-build-OWOug_/wxPython/ext/wxWidgets/src/common/wincmn.cpp
  (3271) in CaptureMouse(): Recapturing the mouse in the same window?

Can be seen in the demo by enabling the EVT_TREE_BEGIN_DRAG checkbox
and trying to drag an item on a Linux/GTK platform.