Add GetFocusedItem method to CustomTreeCtrl
The GetSelection() method of CustomTreeCtrl and HyperTreeList have the
same behavior as the GetFocusedItem() method of wx.TreeCtrl.
The GetCurrentItem() method of HyperTreeList also does the exact same
thing. Try to consolidate and document these methods.

The CustomTreeCtrl "Basic Functionalites" section states that
"All the methods available in TreeCtrl are also available in
CustomTreeCtrl", this brings it one step closer to reality.