Add sort icon support to HyperTreeList
Feature addition.
Add SetColumnSortIcon() method to HyperTreeList so that a native sort
arrow can be easily shown in the header columns. For some reason does
not work on GTK (won't draw arrow) but works great on Windows/MacOS.

Modify the demo with LIST_COL_CLICK event options so it can be tested.