Fix building with Python 3.8 on Linux
In Python 3.8, running "python3-config --libs" no longer returns the python
library itself.  This causes Waf's pyembed check to fail as it can't resolve
references to methods in the python library.  There is discussion[1] ongoing
about adding an "embed" mode to python3-config, but at the moment it is not
implemented.  We can work around this, however, in wxPython, by invoking Waf's
check_python_headers() with just the "pyext" feature as "pyembed" isn't needed
in wxPython (as far as I can tell).

[1] https://bugs.python.org/issue36721