Switch macOS CI to use Python from Azure images
Fixes Python 3.10 and should be faster not having to download & install
Python and pip.

This also removes Python 3.6 from macOS CI - the Azure hosted version of
Python 3.6 has a problem with linker flags (it is setting -stack_size) and
can't compile Python extensions.  Python 3.6 is going to be EOL in a month or
so, so it seems fine to just remove it from CI.