Adds additional bash.exe source.
The way the current build system worked it limited the build to only being able to be performed on a Windows 10+ system when compiling on Windows. This would require a user to have the Unix subsystem feature that is apart of Windows installed. Most users that are going to build wxPython will have Git installed and Git comes with bash.exe so why not use it?