Speeds up Windows build
I am not 100% if this is going to work properly. It does compile correctly and it is substantially faster. for the wxWidgets compilation I use the msvc build system I added to check if cmake and ninja are available and if they are I use that if JOM is not being used. the --jobs command line switch sets the number of parallel files that can be compiled.

Windows builds on Azure typically take around 45 minutes, we will see how much that gets cut down. It took 1 minutes 30 seconds to compile wxWidgets on my 6 core using 6 jobs. I don't remember off hand how long it took previously but I sure know it didn't take a minute and 30 seconds!!