Support newer PyMuPDF versions
PyMuPDF 1.18 renamed all of its camelCase fields and methods to use
snake_case instead, and a later version (probably 1.19) removed the
deprecated aliases. This patch attempts to use the newer name first
(since most people will probably be using a newer version of PyMuPDF)
and falls back to the older name if that fails.