IntCtrl: Change default colour to wx.NullColour
Fixes #2199.

The problem is that the default colour is currently wx.BLACK which will basically be invisible if the user has a dark theme. Use ``wx.NullColour`` instead so that it can just use the default text colour for the user's theme.

Note that this does not fix the second part of issue #2199 (the colours still don't change if the control is disabled) but I feel this should be a separate issue now.