add Window.GetHandleType
This new method returns a string identifying what the
platform-specific return type of Window.GetHandle is. This info can be
sort of guessed from PlatformInfo.PortId, but there is the same issue as
with GetHandle when running on the GTK port: it's not possible to know
without some extra C++ code what the actual windowing backend is.