Fix using `setuptools.build_meta` backend
Add the current directory path to `sys.path` explicitly, in order to fix
using the correct `setuptools.build_meta` backend.  The legacy backend
is only meant to be used implicitly for packages that have not been
ported to PEP 517 builds yet, and using it explicitly is invalid.
The correct solution here is to set the import path correctly.